+++ tools/xdhwebq/htdocs/prolog.php
@@ -36,7 +36,16 @@ $out = '
 		</head>
 		<!--body id="Root" data-xdh-onevents="(keypress|About|SC+a)(keypress|Q37Refresh|SC+r)"-->
 		<body onload="new QRCode(\'qrcode\', {width:125, height:125, correctLevel: QRCode.CorrectLevel.L}).makeCode(window.location.href)">
-			<span>
+			<div style="display:table; margin: 5px auto 10px auto;">
+				<div style="display: table; margin: auto; border: 1px solid; padding: 10px;box-shadow: 5px 5px; text-align: center;">
+					<span style="text-shadow: 1px 1px;">
+						<span>Powered by the <span style="font-style: italic;">Atlas</span> toolkit (<a href="http://atlastk.org/" target="_blank">http://atlastk.org/</a>)</span>
+						<br/>
+						<span>⯈⯈⯈&nbsp;&nbsp;&nbsp;Live demonstrations <a href="https://q37.info/s/sssznrb4" target="_blank">here</a>!&nbsp;&nbsp;&nbsp;⯇⯇⯇</span>
+					</span>
+				</div>
+			</div>
+			<div>
 				<noscript>
 					<div style="display: table; margin: 50px auto auto auto;">
 						<fieldset>
@@ -46,8 +55,8 @@ $out = '
 				</noscript>
 				<input type="hidden" id="LoginUserID" value="' . $UserID . '"/>
 				<input type="hidden" id="LoginPassword" value="' . $Password . '"/>
-			</span>
-			<span>
+			</div>
+			<div>
 				<div style="display:table; margin: 5px auto auto auto;">
 					<hr/>
 					<span style="display: table; margin: 10px auto 5px auto;font-style: oblique;">For another instance, click or scan:</span>
@@ -56,16 +65,16 @@ $out = '
  						   <div id="qrcode"></div>
 						</a>
 					</div>
-						<span style="display: table; margin: auto;font-style: oblique;margin: 5px auto auto auto;">Generated with <a href="https://davidshimjs.github.io/qrcodejs/" target="_blank">qrcode.js</a></span>
-						<div style="display: table; margin: 15px auto 10px auto; border: 1px solid; padding: 10px;box-shadow: 5px 5px; text-align: center;">
+					<span style="display: table; margin: auto;font-style: oblique;margin: 5px auto auto auto;">Generated with <a href="https://davidshimjs.github.io/qrcodejs/" target="_blank">qrcode.js</a></span>
+					<!--div style="display: table; margin: 15px auto 10px auto; border: 1px solid; padding: 10px;box-shadow: 5px 5px; text-align: center;">
 						<span style="text-shadow: 1px 1px;">
 							<span>Powered by the <span style="font-style: italic;">Atlas</span> toolkit (<a href="http://atlastk.org/" target="_blank">http://atlastk.org/</a>)</span>
 							<br/>
 							<span>⯈⯈⯈&nbsp;&nbsp;&nbsp;<a href="https://q37.info/s/sssznrb4" target="_blank">Live demonstrations</a>&nbsp;&nbsp;&nbsp;⯇⯇⯇</span>
 						</span>
-					 </div>
+					 </div-->
 				 </div>
-			</span>
+			</div>
 		</body>
 	</html>'
 ?>