+++ other/libs/tortoise/PYH/main.py
@@ -1,4 +1,40 @@
+# coding: utf-8
+"""
+MIT License
+
+Copyright (c) 2019 Claude SIMON (https://q37.info/s/rmnmqd49)
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
+import os, sys
+sys.path.append("./Atlas.python.zip")
+sys.path.append("./Tortoise.python.zip")
+# sys.path.append("./tortoise")
+
+# For using in Repl.it
+if ('HOME' in os.environ) and (os.environ['HOME'] == '/home/runner'):
+  os.environ["ATK"] = "REPLit"
+
+import atlastk
 from tortoise import *
+
 
 HTML = """
 <div style="display: table; margin: auto;">
@@ -102,7 +138,7 @@ def drawing4(tortoise):
 
 
 def clear(dom):
-    dom.setLayout("SVG", Atlas.createHTML())
+    dom.setLayout("SVG", atlastk.createHTML())
     tortoise = Tortoise(dom)
     tortoise.setAutoDraw(autoDraw)
     return tortoise
@@ -115,12 +151,12 @@ def call(dom, id):
     tortoise.draw()
 
 
-def acConnect(dummy, dom):
+def acConnect(dom):
     dom.setLayout("", HTML)
     dom.enableElements(ids)
 
 
-def acAll(dummy, dom, id):
+def acAll(dom, id):
     call(dom, 0)
     call(dom, 1)
     call(dom, 2)
@@ -130,9 +166,9 @@ def acAll(dummy, dom, id):
     globals()["autoDraw"] = 50
 
 
-def acDraw(dummy, dom, id):
+def acDraw(dom, id):
     call(dom, id)
     dom.enableElements(ids)
 
 
-Atlas.launch({"": acConnect, "All": acAll, "Draw": acDraw}, None, HEAD)
+atlastk.launch({"": acConnect, "All": acAll, "Draw": acDraw}, None, HEAD)