+++ tools/xdhq/examples/PYH/15-puzzle/main.py
@@ -25,8 +25,8 @@ SOFTWARE.
 import sys
 from random import *
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
+sys.path.append("./src")
+sys.path.append("../src")
 
 import atlastk as Atlas