+++ tools/xdhq/examples/PYH/Contacts/main.py
@@ -26,8 +26,8 @@ import os
 import sys
 from enum import Enum, auto
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
+sys.path.append("./src")
+sys.path.append("../src")
 
 import atlastk as Atlas