+++ tools/xdhq/examples/PYH/RGB/main.py
@@ -24,9 +24,9 @@ SOFTWARE.
 
 import GPIOq, sys
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
-sys.path.append("./Atlas")
+sys.path.append("./src")
+sys.path.append("../src")
+# sys.path.append("./Atlas")
 
 import atlastk as Atlas