+++ tools/xdhq/examples/PYH/ReversiIMG/main.py
@@ -32,8 +32,8 @@ import random
 import itertools
 import time
 
-sys.path.append("./Atlas.python.zip")
-sys.path.append("../Atlas.python.zip")
+sys.path.append("./src")
+sys.path.append("../src")
 
 import atlastk as Atlas