+++ tools/xdhq/examples/PYH/Contacts/main.py
@@ -26,8 +26,8 @@ import os
 import sys
 from enum import Enum, auto
 
-sys.path.append("./src")
-sys.path.append("../src")
+sys.path.append("./atlastk")
+sys.path.append("../atlastk")
 
 import atlastk as Atlas