+++ tools/xdhq/examples/PYH/Stars/main.py
@@ -37,9 +37,9 @@ from pygame.locals import *
 from threading import Thread
 import threading
 
-sys.path.append("./src")
-sys.path.append("../src")
-sys.path.append("../../src")
+sys.path.append("./atlastk")
+sys.path.append("../atlastk")
+sys.path.append("../../atlastk")
 
 import atlastk as Atlas