+++ tools/xdhq/Atlas/CHANGELOG.md
@@ -3,6 +3,11 @@
 - Common to **all** bindings,
 - does **not** concern the native code,
 
+## 2019-07-31
+
+- *Python*:
+  - if the `userCallback` parameter of the `Launch(…)` function is or returns `None`, the (optional) first parameter passed to the callbacks will be the *DOM* object (which is otherwise passed as second parameter),
+
 ## 2019-07-28
 
 - *Python*: